requirements.txt
astroid==2.1.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
idna==2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.0
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.4
notebook==5.7.2
numpy==1.15.4
oauthlib==2.1.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
Pygments==2.3.0
pylint==2.2.2
pyparsing==2.3.0
PySocks==1.6.8
python-dateutil==2.7.5
python-pptx==0.6.17
pytz==2018.9
pywinpty==0.5.4
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.20.1
requests-oauthlib==1.0.0
runipy==0.1.5
scikit-learn==0.20.2
scipy==1.2.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
stop-words==2018.7.23
terminado==0.8.1
testpath==0.4.2
textblob==0.15.2
tornado==5.1.1
traitlets==4.3.2
tweepy==3.7.0
typed-ast==1.1.0
urllib3==1.24.1
virtualenv==16.2.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.10.11
XlsxWriter==1.1.2
